test: migrate login unit test to ts
